Move alignment types into layout module
