+++ src/lib.rs
@@ -54,4 +54,3 @@ pub mod macros;
 pub use crate::data::*;
 pub use crate::toolkit::*;
 pub use crate::traits::*;
-pub use kas_text::Align;
