+++ src/prelude.rs
@@ -26,7 +26,7 @@ pub use kas::event::{Event, Handler, Manager, ManagerState, Response, SendEvent,
 #[doc(no_inline)]
 pub use kas::geom::{Coord, Offset, Rect, Size};
 #[doc(no_inline)]
-pub use kas::layout::{AxisInfo, Margins, SizeRules, StretchPolicy};
+pub use kas::layout::{Align, AlignHints, AxisInfo, Margins, SizeRules, StretchPolicy};
 #[doc(no_inline)]
 pub use kas::macros::*;
 #[doc(no_inline)]
@@ -34,7 +34,7 @@ pub use kas::text::AccelString;
 #[doc(no_inline)]
 pub use kas::text::{EditableTextApi, Text, TextApi, TextApiExt};
 #[doc(no_inline)]
-pub use kas::{Align, AlignHints, WidgetId};
+pub use kas::WidgetId;
 #[doc(no_inline)]
 pub use kas::{Boxed, TkAction};
 #[doc(no_inline)]
